+++ ext/NonuniformFFTsAMDGPUExt.jl
@@ -5,18 +5,30 @@ using NonuniformFFTs.Kernels: Kernels
 using AMDGPU
 using AMDGPU.Device: @device_override
 
+# Some empirical observations (on AMD MI210, ROCm 6.0.2):
+#
+# - using Bessel functions defined by AMD (@device_override below) is significantly slower
+#   than calling functions in Bessels.jl (at least on Float64), so we disable the overrides
+#   for now;
+#
+# - direct evaluation of the BackwardsKaiserBesselKernel (sinh) is faster than the
+#   KaiserBesselKernel (Bessel function I₀), so BKB is the default;
+#
+# - in both cases, direct evaluation is significantly slower than fast polynomial
+#   approximation (which is not the case on CUDA / A100), so we use FastApproximation by
+#   default.
+
 # This is currently not wrapped in AMDGPU.jl, probably because besseli0 is not defined by
 # SpecialFunctions.jl either (the more general besseli is defined though).
 # See:
 #  - https://doc-july-17.readthedocs.io/en/latest/ROCm_Compiler_SDK/ocml.html for available functions
 #  - https://github.com/JuliaGPU/AMDGPU.jl/blob/master/src/device/gcn/math.jl for functions wrapped in AMDGPU.jl
-# TODO: not sure this is faster than the "fallback" implementation in Bessels.jl
-@device_override Kernels._besseli0(x::Float64) = ccall("extern __ocml_i0_f64", llvmcall, Cdouble, (Cdouble,), x)
-@device_override Kernels._besseli0(x::Float32) = ccall("extern __ocml_i0_f32", llvmcall, Cfloat, (Cfloat,), x)
+# @device_override Kernels._besseli0(x::Float64) = ccall("extern __ocml_i0_f64", llvmcall, Cdouble, (Cdouble,), x)
+# @device_override Kernels._besseli0(x::Float32) = ccall("extern __ocml_i0_f32", llvmcall, Cfloat, (Cfloat,), x)
 
-# As with CUDA, on AMDGPU the direct evaluation of the KaiserBesselKernel seems to be faster
-# than the BackwardsKaiserBesselKernel.
-NonuniformFFTs.default_kernel(::ROCBackend) = KaiserBesselKernel()
+NonuniformFFTs.default_kernel(::ROCBackend) = BackwardsKaiserBesselKernel()
+
+NonuniformFFTs.default_kernel_evalmode(::ROCBackend) = FastApproximation()
 
 # We want the result of this function to be a compile-time constant to avoid some type
 # instabilities, which is why we hardcode the result even though it could be obtained using