+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [v0.6.2](https://github.com/jipolanco/NonuniformFFTs.jl/releases/tag/v0.6.1) - 2024-11-04
+
 ### Changed
 
 - Improve performance of atomic operations (affecting type-1 transforms) on AMD